Playback: extract provider resolution and the source-fetcher factories
Second slice out of the 4 539-line torrent_playback_service.dart, stacked on
T3's lib/services/torrent_playback/:

  * playback_provider_resolution.dart (PlaybackProviderResolution) —
    configuredProviders, defaultConfiguredProvider and isConfigured (were
    _configuredProviders / _defaultConfiguredProvider / _isConfigured). Prefs
    plus the cloud registry only; the provider-picker DIALOG stays on the
    service and now asks this unit for the list it offers.
  * playback_source_fetchers.dart (PlaybackSourceFetchers) — seriesFetcherFor,
    movieFetcherFor and effectiveFetchProvider (was _effectiveFetchProvider),
    consuming PlaybackSourceSearch and PlaybackCandidateRanking.
  * playback_meta.dart — PlaybackMeta, moved verbatim so the fetcher unit can
    take it without importing the service back. torrent_playback_service.dart
    re-exports it, so every existing importer is unchanged.

Bodies and comments move verbatim. The only edits the move forced: the three
resolution helpers and _effectiveFetchProvider made public, cross-unit calls
qualified with the owning class (PlaybackCandidateRanking.loadLadder /
orderCandidatesForRules / shouldPreflightDirectStream / allowsAddonSearch,
PlaybackProviderResolution.defaultConfiguredProvider), the doc references to
[_pickProvider] / [_defaultConfiguredProvider] reworded to point at the new
names, and one `if (...) return true;` given braces because the longer
qualified condition no longer fits one line. dart format re-wrapped six calls
on the new file. The god file itself is untouched apart from the deleted
hunks, the new imports/export, the removed play_loader_art and source_priority
imports, and the call sites that had to name the new class; seven of those were
re-wrapped by hand to stay inside 80 cols.

Preserved quirks (all pinned): playback configuration is API-key-only for
RD/TB/PM/AD and toggle-only for PikPak; an empty key is not configured; a saved
default naming a no-longer-configured provider is ignored in favour of the
first configured one; effectiveFetchProvider passes any non-placeholder
provider through WITHOUT checking that it is configured, and only null /
'local' / 'stremio_direct' / 'stream' fall back; the provider-free episode
fetch keeps direct-URL rows only while the provider-backed one keeps the
addon's torrent rows too.

Debt: two more one-line @visibleForTesting tear-offs on TorrentPlaybackService
(seriesFetcherFor, movieFetcherFor) join T3's block —
test/quick_play_rules_test.dart addresses seriesFetcherFor through the class
and is outside this lane's owned files, and the origin pin addresses both by
their origin names so it can pass unedited across the move. No lib caller
depends on either; the queued follow-up lane deletes the whole block.

Pinned by test/playback_provider_resolution_origin_pin_test.dart (22 tests),
which passes unedited across the move. Mutation check repeated on the new
units: forcing _configuredProviders to accept every provider and deleting the
addon-direct arm of effectiveFetchProvider each made the pin fail; restored.
