+++ CODEMAP.md
@@ -21,7 +21,7 @@ right code instead of re-discovering it. Flutter app; code under `lib/{screens,s
 | `lib/screens/torbox/torbox_downloads_screen.dart` | 7 069 |
 | `lib/screens/debrid_downloads_screen.dart` | 6 444 |
 | `lib/services/video_player_launcher.dart` | 5 769 |
-| `lib/services/torrent_playback_service.dart` | 4 548 |
+| `lib/services/torrent_playback_service.dart` | 4 205 |
 | `lib/services/remote_control/remote_command_router.dart` | 5 100 |
 
 Sources is an independent library: `lib/screens/search/search_sources.dart` (2 804).
@@ -59,7 +59,7 @@ using `search_content_session.dart` and `search_content_actions.dart` shared wit
 `lib/screens/torbox/torbox_downloads_screen.dart` (7 069) ·
 `lib/screens/debrid_downloads_screen.dart` (6 444) ·
 `lib/services/video_player_launcher.dart` (5 769) ·
-`lib/services/torrent_playback_service.dart` (4 548) ·
+`lib/services/torrent_playback_service.dart` (4 205) ·
 `lib/services/remote_control/remote_command_router.dart` (5 100).
 
 `lib/widgets/initial_setup_flow.dart` is a 4-line export of
@@ -394,8 +394,17 @@ Same plan table also lists (not extra “sites”, but still consumers until T1/
   predicates. `playback_source_search.dart` (`PlaybackSourceSearch`) owns the
   fetchers: `searchSeriesPackSources`, `searchCuratedSources`, the per-engine
   listing/fetch helpers and the private candidate/pack curation. Both are
-  widget-free. `TorrentPlaybackService` keeps `@visibleForTesting` constant
-  tear-offs for the public names until the test suites are repointed.
+  widget-free. `playback_provider_resolution.dart`
+  (`PlaybackProviderResolution`, T4) owns the silent provider choice —
+  `configuredProviders`, `defaultConfiguredProvider`, `isConfigured` — which
+  the god file's provider-picker *dialog* also calls.
+  `playback_source_fetchers.dart` (`PlaybackSourceFetchers`, T4) owns the two
+  in-player “Load more sources” factories `seriesFetcherFor` /
+  `movieFetcherFor` plus `effectiveFetchProvider`. `playback_meta.dart` holds
+  `PlaybackMeta`, which `torrent_playback_service.dart` re-exports so its
+  importers are unchanged. All are widget-free. `TorrentPlaybackService` keeps
+  `@visibleForTesting` constant tear-offs for the public names until the test
+  suites are repointed.
 - **`lib/main.dart`** — app shell + nav branch (TV rail / desktop rail / `MobileFloatingNav`), tab indices.
 
 ## Search, sources & addons
